@@ -8,7 +8,7 @@ import 'package:flutter_ecommerce_app/src/widgets/title_text.dart';
 import 'package:flutter_ecommerce_app/src/widgets/extentions.dart';
 
 class MainPage extends StatefulWidget {
-  MainPage({Key key, this.title}) : super(key: key);
+  MainPage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -140,7 +140,7 @@ class _MainPageState extends State<MainPage> {
                         switchInCurve: Curves.easeInToLinear,
                         switchOutCurve: Curves.easeOutBack,
                         child: isHomePageSelected
-                            ? MyHomePage()
+                            ? MyHomePage(title: '',)
                             : Align(
                                 alignment: Alignment.topCenter,
                                 child: ShoppingCartPage(),
@@ -155,7 +155,7 @@ class _MainPageState extends State<MainPage> {
               bottom: 0,
               right: 0,
               child: CustomBottomNavigationBar(
-                onIconPresedCallback: onBottomIconPressed,
+                onIconPresedCallback: onBottomIconPressed, 
               ),
             )
           ],
