@@ -4,7 +4,7 @@ import 'package:flutter_ecommerce_app/src/widgets/BottomNavigationBar/bottom_cur
 
 class CustomBottomNavigationBar extends StatefulWidget {
   final Function(int) onIconPresedCallback;
-  CustomBottomNavigationBar({Key key, this.onIconPresedCallback})
+  CustomBottomNavigationBar({Key? key, required this.onIconPresedCallback})
       : super(key: key);
 
   @override
@@ -16,8 +16,8 @@ class _CustomBottomNavigationBarState extends State<CustomBottomNavigationBar>
     with TickerProviderStateMixin {
   int _selectedIndex = 0;
 
-  AnimationController _xController;
-  AnimationController _yController;
+  late AnimationController _xController;
+  late AnimationController _yController;
   @override
   void initState() {
     _xController = AnimationController(
