@@ -1,7 +1,7 @@
 class Category{
-  int id ;
-  String name ;
-  String image ;
-  bool isSelected ;
+  int? id ;
+  String? name ;
+  String? image ;
+  bool? isSelected ;
   Category({this.id,this.name,this.isSelected = false,this.image});
 }
