@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class CustomRoute<T> extends MaterialPageRoute<T> {
-  CustomRoute({WidgetBuilder builder, RouteSettings settings})
+  CustomRoute({required WidgetBuilder builder, RouteSettings? settings})
       : super(builder: builder, settings: settings);
   @override
   Widget buildTransitions(BuildContext context, Animation<double> animation,
