@@ -1,25 +1,19 @@
 import 'package:flutter/material.dart';
 
 extension OnPressed on Widget {
-  Widget  ripple(Function onPressed, {BorderRadiusGeometry borderRadius =const BorderRadius.all(Radius.circular(5))}) => Stack(
-      children: <Widget>[
-        this,
-        Positioned(
-          left: 0,
-          right: 0,
-          top: 0,
-          bottom: 0,
-          child: FlatButton(
-            shape: RoundedRectangleBorder(
-              borderRadius: borderRadius
-            ),
-              onPressed: () {
-                if (onPressed != null) {
-                  onPressed();
-                }
-              },
-              child: Container()),
-        )
-      ],
+  Widget ripple(Function()? onPressed, {BorderRadiusGeometry borderRadius = const BorderRadius.all(Radius.circular(5))}) {
+    return Material( // Ganti dengan Material
+      child: Ink(
+        child: InkWell(
+          onTap: () {
+            if (onPressed != null) {
+              onPressed();
+            }
+          },
+          borderRadius: borderRadius as BorderRadius,
+          child: this,
+        ),
+      ),
     );
+  }
 }
